Test support (#7)
* Added the first tests

* Check out selucit version of EpoxyDuino

* Added EpoxyDuino to ignore

* Changed config to return a value

* Added more emacs files to ignore

* Updated tests

* Added MockUTCTime libs

* Added clean test to Makefile

* Added Secure Wifi Clinet

* Trying to use github actions to compile

* Fixed syntax error in github action file

* Disable lint in workflow

* Disable lint in workflow

* Updatede examples executing

* Removed compiler warnings from test build.

* Ignore gcov files.

* Return NULL for network if it cannot be created.

* Setup arduinojson lib forced to arduino string, so it can be used by epoxy.

* Do not try to send if not connected.

* Added more coverage ignore files.

* Shortcut to delete coverage files.

* Backup non-working mock server response.

* Return error if socket read returns error.

* Remove check for -1 in read, as that is how arduino return that it has read 0 bytes...

* Set a fixed timestamp in mockutc time.

* Removed hest string.

* Updated socket mock so it can interpret json, and generate responses depending on the test setting.

* Added value types to test case.

* Disable log in mock as default.

* Added a second control test to callback test.

* Added not connected test.

* Added test of step size, and number values.
Removed compiler warning for unused parameters in callback.

* Added testcase to test delete network and device

* Control should also handle decimal point as report does.

* Added test for decimal points for control.

* Fix copy paste error in xml value.

* Added test of xml value creation.

* Added string to callback test.

* Test fresh connect, do uuids not known by server
Verify that data set in report and control are correct after boot.

* Added test for read and write only values.

* Add identifier to delete device and network, as the server should send these.

* Added mock recieve buffer, and validate testcases for sent/generated data

* Remove generic printf line that should not be there.

* Moved testcases to the extras folder.

* Also update directory for clean-test after folder move

* Moved libraries to the extra folder.

* Updated makefiles for tests.

* Test more edges for device and network to increase coverage.

* Removed compile warning, and removed incorrect create value, copy/paste

* Added test with delete without callback registered.

* Try make setup instead

* Fix folder location for makefile.

* Commit patch file in correct folder.

* Use testcov to start test in github.

* Do not use my own full path...

* Remove unused convertToJson function.

+++ src/WappstoIoT.cpp
@@ -5,14 +5,14 @@
 void generateNewUuid(char *str)
 {
     sprintf(str, "%04x%04x-%04x-%04x-%04x-%04x%04x%04x",
-            random(0xFFFF) & 0xFFFF,
-            random(0xFFFF) & 0xFFFF,
-            random(0xFFFF) & 0xFFFF,
-            random(0xFFFF) & 0xFFFF,
-            random(0xFFFF) & 0xBFFF, // UUID variant = 8 .. b
-            random(0xFFFF) & 0xFFFF,
-            random(0xFFFF) & 0xFFFF,
-            random(0xFFFF) & 0xFFFF);
+            (int)(random(0xFFFF) & 0xFFFF),
+            (int)(random(0xFFFF) & 0xFFFF),
+            (int)(random(0xFFFF) & 0xFFFF),
+            (int)(random(0xFFFF) & 0xFFFF),
+            (int)(random(0xFFFF) & 0xBFFF), // UUID variant = 8 .. b
+            (int)(random(0xFFFF) & 0xFFFF),
+            (int)(random(0xFFFF) & 0xFFFF),
+            (int)(random(0xFFFF) & 0xFFFF));
     // UUID verion 4 = random
     str[14] = '4';
     str[36] = '\0';
